refactor: move logStream API from api/rest --> pkg/server/router
